refactor(session): move --id flow to helper; revert data-loader change
Keep `data-loader` untouched per guideline. Extract all session-id lookup
logic to `src/commands/_session_id.ts`, and wire command to call it. Add
locale normalization in command layer.
